change from WebR to iToL delivery
This change was necessary because ggtree could not be installed through WebR. Although the BiocManager Bioconductor library can be installed, due to the lack of curl() support it cannot validate or download all packages.

Installation through devtools::install() from the ggtree GitHub repository fails for the same reason.

Materials are adapted to use iToL instead.